+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Moved map credits to map column so it don't get hidden by chart panel.
 - Fix support for `initUrls` in `startData.initSources`
+- Propagate `knownContainerUniqueIds` across references and their target.
 - [The next improvement]
 
 #### 8.2.13 - 2022-09-01